creat data folder if not exist

+++ src/main.py
@@ -4,15 +4,17 @@ from text2image import generate_picture
 import os
 
 def delete_all_files_in_data_directory():
-    data_directory = os.path.join(os.getcwd(), 'data')
-    for filename in os.listdir(data_directory):
-        file_path = os.path.join(data_directory, filename)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-                print(f'Deletec: {file_path}')
-        except Exception as e:
-            print(f'Erreur lors de la suppression de {file_path}. Raison: {e}')
+    data_dir = os.path.join(os.getcwd(), 'data')
+    if os.path.exists(data_dir):
+        data_directory = os.path.join(os.getcwd(), 'data')
+        for filename in os.listdir(data_directory):
+            file_path = os.path.join(data_directory, filename)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+                    print(f'Deletec: {file_path}')
+            except Exception as e:
+                print(f'Erreur lors de la suppression de {file_path}. Raison: {e}')
 
 if __name__ == "__main__":
     delete_all_files_in_data_directory()
@@ -20,7 +22,7 @@ if __name__ == "__main__":
     prompt = input("Teach me about ")
     # prompt = "the solar system"
     jsonOut = generate_educational_content(prompt)
-    print(jsonOut["json"]["script"])
+    print(jsonOut["json"])
     for chapter_index, chapter in enumerate(jsonOut["json"]["script"]["chapters"]):
         
         generate_audio(chapter["content"],chapter_index)
